server/requirements.txt
auroraapi==0.1.2
Django==1.11.5
django-braces==1.12.0
django-cors-middleware==1.3.1
django-extensions==1.7.8
django-model-utils==3.0.0
django-oauth-toolkit==1.0.0
djangorestframework==3.6.2

